move getLastDeployDate to dedicated util file
no longer in scope of GitHub
