+++ src/lib/utils/getLastDeployDate.ts
@@ -1,0 +1,3 @@
+import published from "@/data/published.json"
+
+export const getLastDeployDate = () => new Date(published.date).toISOString()
